refactor(markdown-index): extract table configs to dedicated file
